commons: move widely used functions to utils
