+++ image_info/image_info/common/cloud_init.py
@@ -1,62 +1,12 @@
 """
 Configuration files
 """
-import os
-import glob
 import contextlib
 import yaml
 from attr import define
 from image_info.report.common import Common
 
-
-def _read_glob_paths_with_parser(tree, glob_paths, parser_func):
-    """
-    Use 'parser_func' to read all files obtained by using all 'glob_paths'
-    globbing patterns under the 'tree' path.
-
-    The 'glob_paths' is a list string patterns accepted by glob.glob().
-    The 'parser_func' function is expected to take a single string argument
-    containing the absolute path to a configuration file which should be parsed.
-    Its return value can be arbitrary representation of the parsed
-    configuration.
-
-    Returns: dictionary with the keys corresponding to directories, which
-    contain configuration files mathing the provided glob pattern. Value of
-    each key is another dictionary with keys representing each filename and
-    values being the parsed configuration representation as returned by the
-    provided 'parser_func' function.
-
-    An example return value for dracut configuration paths and parser:
-    {
-        "/etc/dracut.conf.d": {
-            "sgdisk.conf": {
-                "install_items": " sgdisk "
-            },
-        },
-        "/usr/lib/dracut/dracut.conf.d": {
-            "xen.conf": {
-                "add_drivers": " xen-netfront xen-blkfront "
-            }
-        }
-    }
-    """
-    result = {}
-
-    for glob_path in glob_paths:
-        glob_path_result = {}
-
-        files = glob.glob(f"{tree}{glob_path}")
-        for file in files:
-            config = parser_func(file)
-            if config:
-                filename = os.path.basename(file)
-                glob_path_result[filename] = config
-
-        if glob_path_result:
-            checked_path = os.path.dirname(glob_path)
-            result[checked_path] = glob_path_result
-
-    return result
+from image_info.utils.files import _read_glob_paths_with_parser
 
 
 @define(slots=False)
